+++ pkg/splunk/enterprise/standalone.go
@@ -49,6 +49,7 @@ func ApplyStandalone(ctx context.Context, client splcommon.ControllerClient, cr 
 		cr.Status.ResourceRevMap = make(map[string]string)
 	}
 	eventPublisher, _ := newK8EventPublisher(client, cr)
+	ctx = context.WithValue(ctx, splcommon.EventPublisherKey, eventPublisher)
 	cr.Kind = "Standalone"
 
 	var err error