+++ test/monitoring_console/monitoring_console_suite_test.go
@@ -44,7 +44,10 @@ func TestBasic(t *testing.T) {
 
 	RegisterFailHandler(Fail)
 
-	RunSpecs(t, "Running "+testSuiteName)
+	sc, _ := GinkgoConfiguration()
+	sc.Timeout = 240 * time.Minute
+
+	RunSpecs(t, "Running "+testSuiteName, sc)
 }
 
 var _ = BeforeSuite(func() {